Split project_cli into atomics, adapter, and recipes modules.
Extract gh_project_adapter, project_atomics, and project_recipes from the
monolithic CLI; keep project_cli as a thin cmd dispatcher with re-exports
so tests and project_outbox late imports stay stable.

Author: Savin Ionuț Răzvan
GitHub-User: @SavinRazvan
